+++ lib/vstutils/src/convert.rs
@@ -1,7 +1,7 @@
-pub fn midi_pitch_to_freq(pitch: u8) -> f64 {
+pub fn midi_pitch_to_freq(pitch: u8) -> f32 {
     const A4_PITCH: i8 = 69;
-    const A4_FREQ: f64 = 440.0;
+    const A4_FREQ: f32 = 440.0;
 
     // Midi notes can be 0-127
-    ((f64::from(pitch as i8 - A4_PITCH)) / 12.).exp2() * A4_FREQ
+    ((f32::from(pitch as i8 - A4_PITCH)) / 12.).exp2() * A4_FREQ
 }
